+++ app/Main.hs
@@ -9,41 +9,31 @@ import Data.Bool
 import Diagrams.Backend.SVG
 import Diagrams.Prelude hiding (rad)
 import Web.Suavemente
-import Web.Suavemente.Orphans
+import Web.Suavemente.Diagrams
 
 
 main :: IO ()
-main = suavementely
-  [ ("zoom", SomeSuave $ do
-      sq  <- dropdown "Shape"
-               [ ("Triangle", "0")
-               , ("Square", "1")
-               , ("Circle", "2")
-               , ("Star", "3")
-               ] ("0" :: String)
-      rad <- slider "Radius" 1 10 5
-      n   <- slider "Polys" 5 12 5
-      r   <- realSlider "Red" 0 1 0.05 0.5
-      g   <- realSlider "Green" 0 1 0.05 1
-      b   <- realSlider "Blue" 0 1 0.05 1
-      x   <- slider "X" 0 20 10
-      y   <- slider "Y" 0 20 10
+main = suavemente sendDiagram $ do
+  sq  <- dropdown "Shape"
+           [ ("Triangle", "0")
+           , ("Square", "1")
+           , ("Circle", "2")
+           , ("Star", "3")
+           ] ("0" :: String)
+  rad <- slider "Radius" 1 10 5
+  n   <- slider "Polys" 5 12 5
+  r   <- realSlider "Red" 0 1 0.05 0.5
+  g   <- realSlider "Green" 0 1 0.05 1
+  b   <- realSlider "Blue" 0 1 0.05 1
+  x   <- slider "X" 0 20 10
+  y   <- slider "Y" 0 20 10
 
-      pure (
-        getShape sq rad n
-                # fc (sRGB r g b)
-                # translate (r2 (x, y))
-                # rectEnvelope (p2 (0, 0)) (r2 (20, 20))
-        :: Diagram B)
-      )
-
-  , ("fwop", SomeSuave $ do
-      bool <- checkbox "Check" False
-      string <- textbox "String" "hello"
-
-      pure (bool, string)
-    )
-  ]
+  pure (
+    getShape sq rad n
+            # fc (sRGB r g b)
+            # translate (r2 (x, y))
+            # rectEnvelope (p2 (0, 0)) (r2 (20, 20))
+    :: Diagram B)
 
 
 getShape :: String -> Double -> Int -> Diagram B